[Instruments] Add ability to load data from flag Data column.
LORIS previously saved instrument data to a "Data" column
in the flag table as a test, this change gives users the
option to experiment with using it. Loading data from the
Data column means that the instrument SQL tables are no
longer required, but should be considered experimental and
not ready for production.

To test this: set the "$jsonData" variable in an instrument
to true, and try saving and loading and generally using the
instrument. There should be no noticeable difference. If
you're feeling lucky, drop the existing table and continue
trying to save and load data from the instrument.

Note that it might be easier to use a new instrument, since
scoring functions in PHP instruments are often intertwined
with database access.